Add authentication coordinator
