+++ Dokan/Dokan/Internal/Coordinator/Coordinator.swift
@@ -36,50 +36,3 @@ extension Coordinator {
         navigationController.dismiss(animated: animated, completion: completion)
     }
 }
-
-// MARK: - AuthenticationCoordinator
-
-protocol AuthenticationCoordinator: AnyObject {
-    func showRegistration()
-    func showVerification()
-    func didVerifyPhoneNumber()
-}
-
-class DefaultAuthenticationCoordinator: Coordinator {
-    var children: [Coordinator] = []
-    let navigationController: UINavigationController
-    private let onVerifyLogin: () -> Void
-
-    init(onVerifyLogin: @escaping () -> Void) {
-        self.onVerifyLogin = onVerifyLogin
-        navigationController = UINavigationController()
-    }
-
-    func start() {
-        let viewModel = LoginViewModel()
-        let viewController = LoginViewController(viewModel: viewModel)
-        viewController.coordinator = self
-        show(viewController: viewController)
-    }
-}
-
-extension DefaultAuthenticationCoordinator: AuthenticationCoordinator {
-
-    func showRegistration() {
-        let viewModel = RegisterViewModel()
-        let viewController = RegisterViewController(viewModel: viewModel)
-        viewController.coordinator = self
-        show(viewController: viewController)
-    }
-
-    func showVerification() {
-        let viewModel = VerificationViewModel()
-        let viewController = VerificationViewController(viewModel: viewModel)
-        viewController.coordinator = self
-        show(viewController: viewController)
-    }
-
-    func didVerifyPhoneNumber() {
-        onVerifyLogin()
-    }
-}
